@@ -1,8 +1,4 @@
-import Vue from 'vue'
-import Updater from '../../containers/Updater.vue'
+import { createApp } from 'vue';
+import Updater from '../../containers/Updater.vue';
 
-Vue.config.productionTip = false
-
-new Vue({
-  render: h => h(Updater),
-}).$mount('#app')
+createApp(Updater).mount('#app');
